fix(dto): Điều chỉnh các dto response và summary trả về theo kiểu enum

+++ src/main/java/com/example/evsalesmanagement/dto/salesdiscountlevel/SalesDiscountLevelRequestDTO.java
@@ -3,26 +3,32 @@ package com.example.evsalesmanagement.dto.salesdiscountlevel;
 import java.math.BigDecimal;
 
 public class SalesDiscountLevelRequestDTO {
-    private Integer salesFrom;
-    private Integer salesTo;
+    private BigDecimal salesFrom;
+    private BigDecimal salesTo;
     private BigDecimal discountPercentage;
-    public Integer getSalesFrom() {
+
+    public BigDecimal getSalesFrom() {
         return salesFrom;
     }
-    public void setSalesFrom(Integer salesFrom) {
+
+    public void setSalesFrom(BigDecimal salesFrom) {
         this.salesFrom = salesFrom;
     }
-    public Integer getSalesTo() {
+
+    public BigDecimal getSalesTo() {
         return salesTo;
     }
-    public void setSalesTo(Integer salesTo) {
+
+    public void setSalesTo(BigDecimal salesTo) {
         this.salesTo = salesTo;
     }
+
     public BigDecimal getDiscountPercentage() {
         return discountPercentage;
     }
+
     public void setDiscountPercentage(BigDecimal discountPercentage) {
         this.discountPercentage = discountPercentage;
     }
-       
+
 }
